port: vgi_rpc.Reflection.v1, and the drift it exposed
Reflection is a co-hosted protocol with list_protocols and describe, registered
after the application protocol so it appears in its own output without being
special-cased and so the primary stays the application protocol.  It is exempt
from the version gate: it is what a version-mismatched client calls to learn
what mismatched.

Python and Go now produce an identical protocol_hash for all 87 conformance
methods.  Getting there meant fixing real, shipped drift that nothing could
detect before, because the old comparison could only compare method names and a
format version:

  - Eight conformance handlers widened their return types.  echo_int8 accepted
    int8 and answered int64; echo_float32 accepted float32 and answered
    float64.  The parameter side already carried the narrowing tag, so the
    methods were asymmetric -- and the conformance assertions never noticed
    because they check the value, not the column type.
  - echo_enum answered plain utf8 where Python answers a dictionary.  Status
    now declares its Arrow type through AnnotatedReturn; without an annotation
    Go infers utf8 from the underlying string kind, which is how they diverged.
  - has_return was true for streams and for void unary methods.  A stream's
    ResultSchema is the empty Protocol-level return, and a void method carries
    a zero-field schema rather than a nil one, so neither the nil check nor the
    method type alone is the question being asked.
  - The primary binding still reported the legacy byte-based hash, which
    digests serialized IPC and so cannot be compared with another port at all.

Two framework gaps closed on the way.  The type mapper had no case for a nested
struct, so a struct inside a list was unrepresentable -- serialization and
deserialization both already handled it, only the schema derivation was
missing.  And the client now stamps the routing key through the shared codec,
so HTTP, TCP and iroh all get it from one place.

+++ conformance/impl.go
@@ -560,27 +560,34 @@ func inspectPoint(_ context.Context, ctx *vgirpc.CallContext, p inspectPointPara
 }
 
 // --- Annotated types ---
+//
+// These return the *narrow* Go type rather than a widened one. The parameter
+// side already carries an `int8`/`float32` tag override, so a widened return
+// made the method asymmetric: it accepted int8 and answered int64. Nothing
+// caught it until protocol_hash became comparable across ports, because the
+// conformance assertions only check the value, and the describe comparison
+// only ever compared method names.
 
-func echoInt32(_ context.Context, ctx *vgirpc.CallContext, p echoInt32Params) (int64, error) {
-	return p.Value, nil
+func echoInt32(_ context.Context, ctx *vgirpc.CallContext, p echoInt32Params) (int32, error) {
+	return int32(p.Value), nil
 }
-func echoFloat32(_ context.Context, ctx *vgirpc.CallContext, p echoFloat32Params) (float64, error) {
-	return p.Value, nil
+func echoFloat32(_ context.Context, ctx *vgirpc.CallContext, p echoFloat32Params) (float32, error) {
+	return float32(p.Value), nil
 }
-func echoInt8(_ context.Context, ctx *vgirpc.CallContext, p echoInt8Params) (int64, error) {
-	return p.Value, nil
+func echoInt8(_ context.Context, ctx *vgirpc.CallContext, p echoInt8Params) (int8, error) {
+	return int8(p.Value), nil
 }
-func echoInt16(_ context.Context, ctx *vgirpc.CallContext, p echoInt16Params) (int64, error) {
-	return p.Value, nil
+func echoInt16(_ context.Context, ctx *vgirpc.CallContext, p echoInt16Params) (int16, error) {
+	return int16(p.Value), nil
 }
-func echoUint8(_ context.Context, ctx *vgirpc.CallContext, p echoUint8Params) (uint64, error) {
-	return p.Value, nil
+func echoUint8(_ context.Context, ctx *vgirpc.CallContext, p echoUint8Params) (uint8, error) {
+	return uint8(p.Value), nil
 }
-func echoUint16(_ context.Context, ctx *vgirpc.CallContext, p echoUint16Params) (uint64, error) {
-	return p.Value, nil
+func echoUint16(_ context.Context, ctx *vgirpc.CallContext, p echoUint16Params) (uint16, error) {
+	return uint16(p.Value), nil
 }
-func echoUint32(_ context.Context, ctx *vgirpc.CallContext, p echoUint32Params) (uint64, error) {
-	return p.Value, nil
+func echoUint32(_ context.Context, ctx *vgirpc.CallContext, p echoUint32Params) (uint32, error) {
+	return uint32(p.Value), nil
 }
 func echoUint64(_ context.Context, ctx *vgirpc.CallContext, p echoUint64Params) (uint64, error) {
 	return p.Value, nil
